split utils file a little
