+++ src/__tests__/utils.js
@@ -11,7 +11,6 @@ import {
     _isBlockedUA,
     DEFAULT_BLOCKED_UA_STRS,
     loadScript,
-    _isUrlMatchingRegex,
     isCrossDomainCookie,
 } from '../utils'
 
@@ -249,28 +248,6 @@ describe('loadScript', () => {
                     []
                 )
             ).toBe(true)
-        })
-    })
-
-    describe('_isUrlMatchingRegex', () => {
-        it('returns false when url does not match regex pattern', () => {
-            // test query params
-            expect(_isUrlMatchingRegex('https://example.com', '(\\?|\\&)(name.*)\\=([^&]+)')).toEqual(false)
-            // incorrect route
-            expect(_isUrlMatchingRegex('https://example.com/something/test', 'example.com/test')).toEqual(false)
-            // incorrect domain
-            expect(_isUrlMatchingRegex('https://example.com', 'anotherone.com')).toEqual(false)
-        })
-
-        it('returns true when url matches regex pattern', () => {
-            // match query params
-            expect(_isUrlMatchingRegex('https://example.com?name=something', '(\\?|\\&)(name.*)\\=([^&]+)')).toEqual(
-                true
-            )
-            // match subdomain wildcard
-            expect(_isUrlMatchingRegex('https://app.example.com', '(.*.)?example.com')).toEqual(true)
-            // match route wildcard
-            expect(_isUrlMatchingRegex('https://example.com/something/test', 'example.com/(.*.)/test')).toEqual(true)
         })
     })
