+++ src/extensions/replay/web-performance.ts
@@ -1,6 +1,7 @@
-import { _isUndefined, isLocalhost, logger } from '../../utils'
+import { _isUndefined, logger } from '../../utils'
 import { PostHog } from '../../posthog-core'
 import { DecideResponse, NetworkRequest } from '../../types'
+import { isLocalhost } from '../../request-utils'
 
 const PERFORMANCE_EVENTS_MAPPING: { [key: string]: number } = {
     // BASE_PERFORMANCE_EVENT_COLUMNS
